docs reword - remove extra 'both' and reword
Reword a clunky sentence by removing an out of place 'both' and trying
to clarify what is meant.

This might not be the right answer -- created a sentence that means you
can access both lockdown and other RemoteXPC services. If this wasn't
the desired meaning maybe it should be "all the other lockdown and
RemoteXPC services" instead.

Also it's possible 'lockdown' should be 'lockdownd' instead.