Close TCP server on exceptions
When running a command like `pymobiledevice3 developer debugserver start-server 3921` with tunneld, it first tries without tunneld, raise an exception, which climbs back up in the call stack, bypassing the closing of the socket, leading the next attempt to fail as the port is set and that the previous server has never been closed.
This commit fixes that issue.